+++ src/methods/forward_message.rs
@@ -9,9 +9,9 @@ pub struct ForwardMessage<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Unique identifier for the chat where the original message was sent (or channel username in the format @channelusername)
-    pub from_chat_id: i32,
+    pub from_chat_id: i64,
     /// Sends the message silently. Users will receive a notification with no sound.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub disable_notification: Option<bool>,
@@ -20,7 +20,7 @@ pub struct ForwardMessage<'a> {
 }
 
 impl<'a> ForwardMessage<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, from_chat_id: i32, message_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, from_chat_id: i64, message_id: i32) -> Self {
         Self {
             chat_id: chat_id,
             from_chat_id: from_chat_id,
@@ -34,11 +34,11 @@ impl<'a> ForwardMessage<'a> {
         let resp = self.bot.send("forwardMessage", Some(string)).await?;
         Ok(serde_json::from_str::<Message>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
-    pub fn from_chat_id(mut self, from_chat_id: i32) -> Self {
+    pub fn from_chat_id(mut self, from_chat_id: i64) -> Self {
         self.from_chat_id = from_chat_id;
         self
     }
