+++ src/methods/kick_chat_member.rs
@@ -9,7 +9,7 @@ pub struct KickChatMember<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target group or username of the target supergroup or channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Unique identifier of the target user
     pub user_id: i32,
     /// Date when the user will be unbanned, unix time. If user is banned for more than 366 days or less than 30 seconds from the current time they are considered to be banned forever. Applied for supergroups and channels only.
@@ -21,7 +21,7 @@ pub struct KickChatMember<'a> {
 }
 
 impl<'a> KickChatMember<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, user_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, user_id: i32) -> Self {
         Self {
             chat_id: chat_id,
             user_id: user_id,
@@ -35,7 +35,7 @@ impl<'a> KickChatMember<'a> {
         let resp = self.bot.send("kickChatMember", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
