+++ src/methods/restrict_chat_member.rs
@@ -9,7 +9,7 @@ pub struct RestrictChatMember<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target supergroup (in the format @supergroupusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Unique identifier of the target user
     pub user_id: i32,
     /// A JSON-serialized object for new user permissions
@@ -20,7 +20,7 @@ pub struct RestrictChatMember<'a> {
 }
 
 impl<'a> RestrictChatMember<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, user_id: i32, permissions: ChatPermissions) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, user_id: i32, permissions: ChatPermissions) -> Self {
         Self {
             chat_id: chat_id,
             user_id: user_id,
@@ -34,7 +34,7 @@ impl<'a> RestrictChatMember<'a> {
         let resp = self.bot.send("restrictChatMember", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
