+++ src/methods/send_audio.rs
@@ -9,7 +9,7 @@ pub struct SendAudio<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Audio file to send. Pass a file_id as String to send an audio file that exists on the Telegram servers (recommended), pass an HTTP URL as a String for Telegram to get an audio file from the Internet, or upload a new one using multipart/form-data. More info on Sending Files
     pub audio: InputFile,
     /// Audio caption, 0-1024 characters after entities parsing
@@ -48,7 +48,7 @@ pub struct SendAudio<'a> {
 }
 
 impl<'a> SendAudio<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, audio: InputFile) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, audio: InputFile) -> Self {
         Self {
             chat_id: chat_id,
             audio: audio,
@@ -71,7 +71,7 @@ impl<'a> SendAudio<'a> {
         let resp = self.bot.send("sendAudio", Some(string)).await?;
         Ok(serde_json::from_str::<Message>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
