+++ src/methods/send_invoice.rs
@@ -9,7 +9,7 @@ pub struct SendInvoice<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Product name, 1-32 characters
     pub title: String,
     /// Product description, 1-255 characters
@@ -84,7 +84,7 @@ pub struct SendInvoice<'a> {
 impl<'a> SendInvoice<'a> {
     pub fn new(
         bot: &'a Bot,
-        chat_id: i32,
+        chat_id: i64,
         title: String,
         description: String,
         payload: String,
@@ -127,7 +127,7 @@ impl<'a> SendInvoice<'a> {
         let resp = self.bot.send("sendInvoice", Some(string)).await?;
         Ok(serde_json::from_str::<Message>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
