+++ src/methods/send_photo.rs
@@ -9,7 +9,7 @@ pub struct SendPhoto<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Photo to send. Pass a file_id as String to send a photo that exists on the Telegram servers (recommended), pass an HTTP URL as a String for Telegram to get a photo from the Internet, or upload a new photo using multipart/form-data. The photo must be at most 10 MB in size. The photo's width and height must not exceed 10000 in total. Width and height ratio must be at most 20. More info on Sending Files
     pub photo: InputFile,
     /// Photo caption (may also be used when resending photos by file_id), 0-1024 characters after entities parsing
@@ -36,7 +36,7 @@ pub struct SendPhoto<'a> {
 }
 
 impl<'a> SendPhoto<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, photo: InputFile) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, photo: InputFile) -> Self {
         Self {
             chat_id: chat_id,
             photo: photo,
@@ -55,7 +55,7 @@ impl<'a> SendPhoto<'a> {
         let resp = self.bot.send("sendPhoto", Some(string)).await?;
         Ok(serde_json::from_str::<Message>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
