+++ src/methods/set_chat_photo.rs
@@ -9,13 +9,13 @@ pub struct SetChatPhoto<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// New chat photo, uploaded using multipart/form-data
     pub photo: InputFile,
 }
 
 impl<'a> SetChatPhoto<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32, photo: InputFile) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64, photo: InputFile) -> Self {
         Self {
             chat_id: chat_id,
             photo: photo,
@@ -27,7 +27,7 @@ impl<'a> SetChatPhoto<'a> {
         let resp = self.bot.send("setChatPhoto", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
