+++ src/methods/unpin_all_chat_messages.rs
@@ -9,11 +9,11 @@ pub struct UnpinAllChatMessages<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
 }
 
 impl<'a> UnpinAllChatMessages<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64) -> Self {
         Self {
             chat_id: chat_id,
             bot: bot,
@@ -24,7 +24,7 @@ impl<'a> UnpinAllChatMessages<'a> {
         let resp = self.bot.send("unpinAllChatMessages", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
