+++ src/methods/unpin_chat_message.rs
@@ -9,14 +9,14 @@ pub struct UnpinChatMessage<'a> {
     #[serde(skip)]
     bot: &'a Bot,
     /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
-    pub chat_id: i32,
+    pub chat_id: i64,
     /// Identifier of a message to unpin. If not specified, the most recent pinned message (by sending date) will be unpinned.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub message_id: Option<i32>,
 }
 
 impl<'a> UnpinChatMessage<'a> {
-    pub fn new(bot: &'a Bot, chat_id: i32) -> Self {
+    pub fn new(bot: &'a Bot, chat_id: i64) -> Self {
         Self {
             chat_id: chat_id,
             message_id: None,
@@ -28,7 +28,7 @@ impl<'a> UnpinChatMessage<'a> {
         let resp = self.bot.send("unpinChatMessage", Some(string)).await?;
         Ok(serde_json::from_str::<bool>(&resp.text().await?)?)
     }
-    pub fn chat_id(mut self, chat_id: i32) -> Self {
+    pub fn chat_id(mut self, chat_id: i64) -> Self {
         self.chat_id = chat_id;
         self
     }
